pyproject.toml
[tool.poetry]
name = "humanizer"
version = "1.0.0"
description = "Transformation via Recursive Measurement - Consciousness work through quantum-inspired ML"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "humanizer"}]

[tool.poetry.dependencies]
python = "^3.11"
# API
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
python-multipart = "^0.0.6"

# Database
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"
pgvector = "^0.2.3"

# ML / TRM
torch = "^2.1.0"
numpy = "^1.26.2"
sentence-transformers = "^2.2.2"
scikit-learn = "^1.3.2"

# Utilities
python-dotenv = "^1.0.0"
httpx = "^0.25.2"
pydantic-settings = "^2.1.0"

# Document Processing
pypdf2 = "^3.0.0"
pdfplumber = "^0.11.0"
chardet = "^5.2.0"
pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.12.0"
ruff = "^0.1.7"
mypy = "^1.7.1"
ipython = "^8.18.1"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
